+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - static getCacheKey function in the CacheInterceptor to get the key of a given translation.
 - [Issue 1](https://github.com/tobytwigger/laravel-translate/issues/1) - Clear the cache when a database translation is saved.
 - [Issue 10](https://github.com/tobytwigger/laravel-translate/issues/10) - Changed the _translate API to a get request for proper caching.
+- [Issue 7](https://github.com/tobytwigger/laravel-translate/issues/7) - Added supported languages key to config.
 
 ## [0.1] - (09/09/2020)